Added modals for Login, linked up & working
