+++ src/App.js
@@ -1,17 +1,21 @@
 /* jshint esversion: 6*/
 import React, { Component } from 'react';
 
-import { LoginModal } from './components/Exports';
-
 //Custom imports
 import { MenuBar, CustomSideBar } from './components/Exports';
+
+//Modals
+import { LoginModal } from './components/Modals';
+
+//Config
+import { USER_COOKIE_IDENTIFIER } from './config';
 
 import './App.css';
 
 class App extends Component {
   constructor(props) {
     super(props);
-    this.state = { visible: false }
+    this.state = { visible: false };
     this.toggleVisibility = this.toggleVisibility.bind(this);
   }
 
@@ -20,7 +24,7 @@ class App extends Component {
   }
 
   getSession() {
-        return window.localStorage ? localStorage.getItem("token") !== null : null;
+        return window.localStorage ? localStorage.getItem(USER_COOKIE_IDENTIFIER) === null : null;
   } 
 
   render() {   
@@ -28,7 +32,7 @@ class App extends Component {
           <div className = "App">
               <MenuBar toggleVisibility = {this.toggleVisibility} />
               <CustomSideBar visible = {this.state.visible} />  
-              <LoginModal open = {!this.getSession()}/> 
+              <LoginModal open = {this.getSession()}/> 
           </div>
       );
   }
