--- src/components/Modals/Exports/index.js
@@ -1,18 +1,0 @@
-/* jshint esversion: 6 */
-import SemanticModal from '../Template';
-
-//User Settings
-import UserSettingsModal from '../UserSettings';
-
-//Confirm Modal
-import ConfirmModal from '../Confirm';
-
-//Hub Settings
-import HubSettingsModal from '../HubSettings';
-
-export {
-    SemanticModal,
-    UserSettingsModal,
-    ConfirmModal,
-    HubSettingsModal
-};
