+++ src/components/Modals/index.js
@@ -1,0 +1,14 @@
+/* jshint esversion: 6 */
+import SemanticModal from './Template';
+import UserSettingsModal from './UserSettings';
+import ConfirmModal from './Confirm';
+import HubSettingsModal from './HubSettings';
+import LoginModal from './Login';
+
+export {
+    SemanticModal,
+    UserSettingsModal,
+    ConfirmModal,
+    HubSettingsModal,
+    LoginModal
+};
